# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: MockAgents API
  version: 0.4.0
  description: >
    MockAgents is a platform for mocking AI agent integrations. It exposes a
    Management API (agents, interaction logs, costs, audit, pipelines, config
    validation, and a multi-tenant control plane of tenants + RBAC API keys),
    and Protocol Endpoints that mimic real LLM provider APIs (OpenAI Chat
    Completions and Anthropic Messages).


    The Model Context Protocol mock (`/mcp`, `/mcp/events`, `/mcp/response`) is
    JSON-RPC 2.0 rather than REST and is documented in the project README,
    not in this OpenAPI document.


    **Authentication.** When the server runs with `MOCKAGENTS_MULTI_TENANT=1`,
    every `/api/v1/*` management route requires an API key
    (`Authorization: Bearer <key>` or `X-Api-Key: <key>`) whose role meets the
    route's floor (`viewer < editor < admin < platform`). In single-tenant mode
    the management API is open. The protocol endpoints (`/v1/*`) are always
    unauthenticated — clients send their own provider key, which MockAgents
    ignores.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/agents:
    get:
      operationId: listAgents
      summary: List all loaded agents
      description: Returns a summary array of every agent currently loaded in the server.
      tags: [Agents]
      responses:
        "200":
          description: List of agent summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentSummary"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createAgent
      summary: Create a new agent at runtime
      description: >
        Registers a new agent from a YAML or JSON `Agent` definition. The agent
        serves immediately (no restart) and is persisted to the agents directory
        when one is configured. The body is validated with the same rules as the
        CLI; an invalid body returns 422 and changes nothing. In multi-tenant
        mode the agent is owned by the caller's tenant and the route requires the
        `editor` role. Emits an `agent.created` audit event.
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/yaml:
            schema: { type: string }
          application/json:
            schema: { $ref: "#/components/schemas/AgentDefinition" }
      responses:
        "201":
          description: Agent created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentWriteResult" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: An agent of that name already exists (use PUT to replace).
        "422":
          description: Validation failed; the body lists the errors.
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/agents/{name}:
    get:
      operationId: getAgent
      summary: Get full agent definition
      description: >
        Returns the COMPLETE parsed agent configuration — every server-supported
        field, not the reduced summary shape used by the list route — so an
        editor can round-trip it safely. The response carries an `ETag` plus the
        two revision headers described below.
      tags: [Agents]
      parameters:
        - name: name
          in: path
          required: true
          description: Unique agent name.
          schema:
            type: string
      responses:
        "200":
          description: Full agent definition.
          headers:
            ETag:
              description: >
                Opaque revision of this agent. Echo it back as `If-Match` on a
                later PUT to make the write conditional. It covers BOTH the
                running definition and its backing file, so it changes if
                either moves.
              schema: { type: string }
            X-Mockagents-Revision-Effective:
              description: Hash of the definition the engine is currently serving.
              schema: { type: string }
            X-Mockagents-Revision-Source:
              description: >
                Hash of the backing YAML file's bytes. Absent when the agent has
                no file (runtime-only) or the file could not be read. When this
                differs from the effective revision, the file has been edited
                out of band and the running definition is older than disk.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDefinition"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      operationId: putAgent
      summary: Create or replace an agent at runtime
      description: >
        Create-or-replace (upsert). The path name is authoritative and the
        body's `metadata.name` must match it. The change is live immediately and
        persisted when an agents directory is configured. Returns 201 when a new
        agent was created or 200 when an existing one was replaced. Requires the
        `editor` role in multi-tenant mode. Emits `agent.created` or
        `agent.updated`.


        **Conditional writes are additive.** A request that sends neither
        `If-Match` nor `If-None-Match` behaves exactly as it always has:
        unconditional create-or-replace. Existing clients are unaffected.


        **Strict fields ride along with the opt-in.** A request carrying either
        precondition also gets unknown fields REJECTED (422) instead of silently
        dropped. Without this, an editor doing GET → modify → PUT would delete
        any field its types do not model, while the response still said
        `"persisted": true`. An unconditional PUT keeps the older lenient
        behaviour, where unknown fields are discarded.


        **Known limit:** because the route stays additive, a client that sends
        no precondition still overwrites unconditionally. Conditional writes
        order only the callers that use them — they cannot stop a legacy writer.
        A conditional writer will at least be told (412) rather than silently
        losing its work.
      tags: [Agents]
      parameters:
        - name: name
          in: path
          required: true
          description: Agent name (must equal the body's metadata.name).
          schema: { type: string }
        - name: If-Match
          in: header
          required: false
          description: >
            Apply only if the agent's current revision matches. Send the `ETag`
            from a prior GET to replace a known revision, or `*` to require that
            the agent merely exists. A mismatch returns 412 and changes nothing.
            Fails if the agent was deleted, replaced by another writer, or its
            YAML was edited on disk since it was read.
          schema: { type: string }
        - name: If-None-Match
          in: header
          required: false
          description: >
            Only `*` is accepted: create-only. Returns 412 rather than
            overwriting an agent that already exists. Any other value is
            rejected rather than guessed at.
          schema: { type: string, enum: ["*"] }
      requestBody:
        required: true
        content:
          application/yaml:
            schema: { type: string }
          application/json:
            schema: { $ref: "#/components/schemas/AgentDefinition" }
      responses:
        "200":
          description: Existing agent replaced.
          headers:
            ETag:
              description: >
                Opaque revision of this agent. Echo it back as `If-Match` on a
                later PUT to make the write conditional. It covers BOTH the
                running definition and its backing file, so it changes if
                either moves.
              schema: { type: string }
            X-Mockagents-Revision-Effective:
              description: Hash of the definition the engine is currently serving.
              schema: { type: string }
            X-Mockagents-Revision-Source:
              description: >
                Hash of the backing YAML file's bytes. Absent when the agent has
                no file (runtime-only) or the file could not be read. When this
                differs from the effective revision, the file has been edited
                out of band and the running definition is older than disk.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentWriteResult" }
        "201":
          description: New agent created.
          headers:
            ETag:
              description: >
                Opaque revision of this agent. Echo it back as `If-Match` on a
                later PUT to make the write conditional. It covers BOTH the
                running definition and its backing file, so it changes if
                either moves.
              schema: { type: string }
            X-Mockagents-Revision-Effective:
              description: Hash of the definition the engine is currently serving.
              schema: { type: string }
            X-Mockagents-Revision-Source:
              description: >
                Hash of the backing YAML file's bytes. Absent when the agent has
                no file (runtime-only) or the file could not be read. When this
                differs from the effective revision, the file has been edited
                out of band and the running definition is older than disk.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentWriteResult" }
        "412":
          description: >
            A precondition failed and NOTHING was changed. The body explains
            which revision moved — another writer, a deletion, or an out-of-band
            edit to the YAML on disk — and the response carries the agent's
            current `ETag` so a client can rebase without a second round-trip.
          headers:
            ETag:
              description: >
                Opaque revision of this agent. Echo it back as `If-Match` on a
                later PUT to make the write conditional. It covers BOTH the
                running definition and its backing file, so it changes if
                either moves.
              schema: { type: string }
            X-Mockagents-Revision-Effective:
              description: Hash of the definition the engine is currently serving.
              schema: { type: string }
            X-Mockagents-Revision-Source:
              description: >
                Hash of the backing YAML file's bytes. Absent when the agent has
                no file (runtime-only) or the file could not be read. When this
                differs from the effective revision, the file has been edited
                out of band and the running definition is older than disk.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: >
            Validation failed and NOTHING was changed; the body lists the
            errors. When a precondition was sent this also covers unsupported
            fields, each reported with its line so an editor can highlight it.
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deleteAgent
      summary: Delete an agent at runtime
      description: >
        Removes the agent from the engine (it stops serving immediately) and
        deletes its persisted file when one exists. Tenant-scoped: a caller can
        only delete an agent owned by its own tenant. Requires the `editor` role
        in multi-tenant mode. Emits an `agent.deleted` audit event.
      tags: [Agents]
      parameters:
        - name: name
          in: path
          required: true
          description: Agent name.
          schema: { type: string }
      responses:
        "200":
          description: Agent deleted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentWriteResult" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/agents/{name}/reload:
    post:
      operationId: reloadAgent
      summary: Reload an agent definition from disk
      description: >
        Re-reads the named agent's definition file from disk and reloads it into
        the server. Emits an `agent.reloaded` audit event. Requires the `viewer`
        role in multi-tenant mode.
      tags: [Agents]
      parameters:
        - name: name
          in: path
          required: true
          description: Agent name.
          schema:
            type: string
      responses:
        "200":
          description: Reload completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReloadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  schemas:
...
    AgentWriteResult:
      type: object
      required: [status, agent, persisted]
      properties:
        status:
          type: string
          enum: [created, updated, deleted]
        agent:
          type: string
          example: my-agent
        persisted:
          type: boolean
          description: Whether the definition was written to the agents directory.
        file:
          type: string
          description: Base filename written/removed (omitted when not persisted).
          example: my-agent.yaml
        revision:
          type: string
          description: >
            The agent's opaque revision AFTER this write, so a client can keep
            editing conditionally without re-fetching. Omitted if it could not
            be computed — an unknown revision is reported as absent, never as a
            stale value.
    AgentSummary:
      type: object
      required: [name, model, protocol, scenario_count, tool_count, persistence]
      properties:
        name:
          type: string
          example: weather-agent
        description:
          type: string
          example: A mock weather-lookup agent.
        model:
          type: string
          description: The model name this mock reports. It is never called.
          example: gpt-4o
        protocol:
          type: string
          enum:
            - openai-chat-completions
            - anthropic-messages
            - google-gemini
            - ollama-chat
            - bedrock-converse
        tags:
          type: array
          items:
            type: string
          example: [weather, demo]
        scenario_count:
          type: integer
          example: 3
        tool_count:
          type: integer
          example: 2
        effective_revision:
          type: string
          description: >
            Revision of the RUNNING definition — the same value the per-agent
            route publishes as `X-Mockagents-Revision-Effective`. Listing it
            avoids a request per agent when comparing what has moved.

            NOT the ETag, and not usable as an `If-Match` value: the ETag also
            covers the backing file, so computing it for a listing would mean
            reading every agent's file. Fetch the agent when you need a
            precondition. Omitted when it could not be computed, which is not
            the same as unchanged.
          example: 8a8c8e24465ea33da5ed64ecfaba62e62e7d3151b344d56768e243bcbea56c13
        persistence:
          type: string
          description: >
            Whether this definition survives a restart. `missing` is
            deliberately distinct from `runtime`: one is a choice, the other is
            a backing file that has gone away while the definition keeps
            serving from memory.
          enum:
            - file      # backed by a file that is present now
            - runtime   # no backing file; created at runtime, lost on restart
            - missing   # a backing file is tracked but is not there any more
          example: file
        file:
          type: string
          description: >
            Base name of the backing file, when there is one. The absolute path
            is server-side detail and is never returned.
          example: weather-agent.yaml
    AgentDefinition:
      type: object
      description: >
        An agent document, in the same envelope the YAML uses — this is what
        `GET /api/v1/agents/{name}` returns and what the write endpoints accept.

        The properties below are the envelope only. What is valid INSIDE `spec`
        is defined by schema/mockagents-v1-agent.json, which the server
        validates against; duplicating it here would create a second definition
        to keep in sync, and the one that drifted would be this one.
      required: [apiVersion, kind, metadata, spec]
      properties:
        apiVersion:
          type: string
          example: mockagents/v1
        kind:
          type: string
          enum: [Agent]
        metadata:
          type: object
          description: Name, description and tags. See the JSON schema for the full shape.
          additionalProperties: true
        spec:
          type: object
          description: >
            Protocol, model, behavior (scenarios), tools, chaos and streaming.
            See schema/mockagents-v1-agent.json.
          additionalProperties: true
...
    ReloadResponse:
      type: object
      required: [reloaded_count, errors]
      properties:
        reloaded_count:
          type: integer
          example: 4
        errors:
          type: array
          items:
            type: string
          example: []

    # -- Error Schema ---------------------------------------------------------
...
  responses:
    BadRequest:
      description: The request was malformed or contained invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: bad_request
              message: "Invalid value for parameter 'limit': must be a positive integer."
              details: {}
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: The requested resource was not found.
              details: {}
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal_error
              message: An unexpected error occurred.
              details: {}
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <api-key>`. Required on `/api/v1/*` routes when
        the server runs with `MOCKAGENTS_MULTI_TENANT=1`.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: "Alternative to `Authorization: Bearer` for the same API key."

# Applies to the management API in multi-tenant mode. Protocol endpoints
# (/v1/*) are always unauthenticated and override this with `security: []`.
security:
  - BearerAuth: []
  - ApiKeyAuth: []